Changes made during meeting 9/29/24
