+++ Models/Enums/Roles.cs
@@ -1,0 +1,10 @@
+﻿namespace ChemStoreWebApp.Models.Enums
+{
+    public enum Roles
+    {
+        Admin,
+        Employee,
+        Associate,
+        Member
+    }
+}
