+++ Pages/Admin/Accounts/Edit.cshtml
@@ -1,4 +1,5 @@
 ﻿@page
+@using ChemStoreWebApp.Models.Enums
 @model ChemStoreWebApp.Pages.Admin.Account.EditModel
 
 @{
