+++ Pages/Admin/Locations/Create.cshtml
@@ -14,7 +14,7 @@
 </head>
 <body>
 
-<h2>Create Location</h2>
+@* <h2>Create Location</h2>
 <hr />
 <div class="row">
     <div class="col-md-4">
@@ -38,7 +38,7 @@
             </div>
         </form>
     </div>
-</div>
+</div> *@
 
 <div>
     <a asp-page="Index">Back to List</a>
