+++ Pages/Admin/Locations/Edit.cshtml
@@ -21,7 +21,7 @@
     <div class="col-md-4">
         <form method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <input type="hidden" asp-for="Location.RoomId" />
+            @* <input type="hidden" asp-for="Location.RoomId" />
                 <div class="mb-3">
                     <label asp-for="Location.RoomNumber" class="control-label"></label>
                     <input asp-for="Location.RoomNumber" class="form-control" style="width:150px; border-radius:5px; height:25px;" />
@@ -37,7 +37,7 @@
                 </div>
             <div class="form-group">
                 <input type="submit" value="Save" class="btn btn-primary" />
-            </div>
+            </div> *@
         </form>
     </div>
 </div>
