+++ Pages/Log.cshtml
@@ -1,5 +1,6 @@
 ﻿@page
 @model ChemStoreWebApp.Pages.LogModel
+@using ChemStoreWebApp.Models.Enums
 @using ChemStoreWebApp.Utilities
 
 @{
