+++ Pages/Shared/_Layout.cshtml
@@ -1,4 +1,5 @@
-﻿<!DOCTYPE html>
+﻿@using ChemStoreWebApp.Models.Enums
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="utf-8" />
