+++ Pages/_AddModal.cshtml
@@ -1,25 +1,26 @@
-﻿@model ChemStoreWebApp.ViewModels.ContainerViewModel
+﻿@using ChemStoreWebApp.Models.Enums
+@model ChemStoreWebApp.ViewModels.ContainerViewModel
 
 <link rel="stylesheet" href="/css/addEditModal.css" asp-append-version="true" />
 
 <form method="post" id="form1" class="modalForm" x-data="{joe:null}">
-    <input type="text" class="filterInput" name="Chemical Name" placeholder="Chemical Name" value=@Model.ChemicalName /> <br />
-    <input type="text" class="filterInput" name="CAS Number" placeholder="CAS Number" value=@Model.CasNumber /> <br />
-    <input type="text" class="filterInput" id="addSupervisor" name="Supervisor" placeholder="Supervisor" value=@Model.Supervisor /> <br />
+    <input type="text" class="filterInput" name="Chemical Name" placeholder="Product Name" value=@Model.ProductName /> <br />
+    @* <input type="text" class="filterInput" name="CAS Number" placeholder="CAS Number" value=@Model.CasNumber /> <br />
+    <input type="text" class="filterInput" id="addSupervisor" name="Supervisor" placeholder="Supervisor" value=@Model.Supervisor /> <br /> *@
 
 
 
-    <select asp-for="Unit" class="filterInput unitSelector" asp-items="Html.GetEnumSelectList<Units>()">
+    @* <select asp-for="Unit" class="filterInput unitSelector" asp-items="Html.GetEnumSelectList<Units>()">
         <option value="" disabled selected>Units</option>
-    </select> <br />
+    </select> <br /> *@
 
     
 
-
+@* 
     <input type="text" class="filterInput" name="Amount" placeholder="Amount" value=@Model.Amount /> <br />
     <select asp-for="BuildingIndex" class="filterInput buildingSelector" asp-items="Html.GetEnumSelectList<Buildings>()" onchange="getRoomsOfBuilding()">
         <option value="" disabled selected hidden>Building</option>
-    </select>
+    </select> *@
     <select asp-for="RoomIndex" class="filterInput roomSelector">
         <option value="" disabled selected hidden>Room Number</option>
     </select> <br />
