+++ Pages/_EditModal.cshtml
@@ -3,14 +3,14 @@
 <link rel="stylesheet" href="/css/addEditModal.css" asp-append-version="true"/>
 
 <form method="post" id="form1" class="modalForm">
-    <input type="text" class="filterInput" id="editChemName" name="Chemical Name" placeholder="Chemical Name" value=@Model.ChemicalName /> <br />
-    <input type="text" class="filterInput" id="editChemCas" name="CAS Number" placeholder="CAS Number" value=@Model.CasNumber /> <br />
+    <input type="text" class="filterInput" id="editChemName" name="Chemical Name" placeholder="Product Name" value=@Model.ProductName /> <br />
+    @* <input type="text" class="filterInput" id="editChemCas" name="CAS Number" placeholder="CAS Number" value=@Model.CasNumber /> <br />
     <input type="text" class="filterInput" id="editChemSup" name="Supervisor" placeholder="Supervisor" value=@Model.Supervisor /> <br />
-    <input type="text" class="filterInput" id="editChemSize" name="Amount" placeholder="Amount" value=@Model.Amount /> <br />
-    <input type="hidden" id="ContainerID" name="ContainerID" value=@Model.ContainerId />
-    <select asp-for="BuildingIndex" class="filterInput buildingSelector" asp-items="Html.GetEnumSelectList<Buildings>()" onchange="getEditRoomsOfBuilding(1)">
+    <input type="text" class="filterInput" id="editChemSize" name="Amount" placeholder="Amount" value=@Model.Amount /> <br /> *@
+    <input type="hidden" id="ContainerID" name="ContainerID" value=@Model.ContainerID />
+    @* <select asp-for="BuildingIndex" class="filterInput buildingSelector" asp-items="Html.GetEnumSelectList<Buildings>()" onchange="getEditRoomsOfBuilding(1)">
         <option value="" selected disabled>Building</option>
-    </select>
+    </select> *@
     <select asp-for="RoomIndex" class="filterInput roomSelector">
         <option value="" selected disabled hidden>Room Number</option> <!--Doesn't display. I don't know why.-->
     </select> <br />
